wacom: Reinstaurate calibration
The "area" setting has a different treatment in the gsettings-desktop-schemas
tablet schema, the 4 double values express the padding (in unitless 0..1
range) on each of the sides of the tablet. It's been done so we don't rely
on input/output units, which we might have not the luxury to access.

Besides that, the dependency on GsdWacomDevice has been cleared.

+++ panels/wacom/calibrator/calibrator-gui.h
@@ -29,10 +29,10 @@
 /* struct to hold min/max info of the X and Y axis */
 typedef struct
 {
-	int x_min;
-	int x_max;
-	int y_min;
-	int y_max;
+	gdouble x_min;
+	gdouble x_max;
+	gdouble y_min;
+	gdouble y_max;
 } XYinfo;
 
 typedef struct CalibArea CalibArea;
